+++ pyboat/ui/util.py
@@ -115,9 +115,9 @@ def load_data(dir_path="./", debug=False, **kwargs):
 
     """
 
-    err_msg1 = "Non-numeric values encountered in\n\n"
-    err_msg2 = "Parsing errors encountered in\n\n"
-    err_suffix = "\n\ncheck input..!"
+    err_msg1 = "Non-numeric values encountered in\n"
+    err_msg2 = "Parsing errors encountered in\n"
+    err_suffix = "\ncheck input file..!"
 
     # returns a list with 1 element, stand alone File Dialog
     file_names = QFileDialog.getOpenFileName(
@@ -188,6 +188,18 @@ def load_data(dir_path="./", debug=False, **kwargs):
             print("Error loading data..")
             return None, f"{err_msg1}{file_name}{err_suffix}", new_dir_path
 
+        # check that we have a standard RangeIndex,
+        # things like MultiIndex point to a faulty header
+        # (to few/to many columns)
+        if debug:
+            print("df columns:", raw_df.columns)
+            print("df shape:", raw_df.shape)
+            print("got df index:", type(raw_df.index))
+        if isinstance(raw_df.index, pd.core.indexes.multi.MultiIndex):
+            msg = (f"Can not parse table header in\n{file_name}\n"
+                   "check number of columns vs. available column names!")
+            return None, msg, new_dir_path
+
         san_df = sanitize_df(raw_df, debug)
         if san_df is None:
             print("Error sanitizing data..")
@@ -222,7 +234,20 @@ def sanitize_df(raw_df, debug=False):
     if debug:
         print("raw columns:", raw_df.columns)
 
-    # return data and empty error message
+    if not isinstance(raw_df.index, pd.core.indexes.range.RangeIndex):
+        msgBox = QMessageBox()
+        msgBox.setWindowTitle("Data Import Warning")
+        msgBox.setIcon(QMessageBox.Warning)
+        msg = ("Found one additional column which will be ignored.\n"
+               "pyBOAT creates its own time axis on the fly\n"
+               "by setting the `Sampling Interval`!")
+        msgBox.setText(msg)
+        msgBox.exec()
+
+        # revert to default integer range index
+        raw_df.reset_index(drop=True, inplace=True)
+
+    # return data
     return raw_df