+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+### pyBOAT 0.9.11
+
+##### Changes
+- table importing checks number of data columns vs. available header entries and raises Warnings/Errors accordingly
+- Wavelet analysis parameters automatically change upon changing the sampling interval (sensitive dynamic defaults)
+
+##### Fixes
+- added `openpyxl` dependency to re-allow `.xlsx` data imports
+- improved status tips for the settings menu
+
 ### pyBOAT 0.9.10
 
 ##### New